Cleaned ga2/gaDisc as much as I could

+++ ga2/gaDisc.py
@@ -1,7 +1,9 @@
-# This program impliments the discrete genetic algorithm
-# here, Discrete means that the new generation is generated
-# all at the same time by considering the top agents at that
-# given point.
+"""
+This program impliments the discrete genetic algorithm
+here, Discrete means that the new generation is generated
+all at the same time by considering the top agents at that
+given point.
+"""
 
 import os
 import ga2
@@ -14,12 +16,13 @@ class Session:
     This is the session class which is used to access all
     the agents in the current session, invoke different
     functions on this session and setting session
-    parameters.
+    parameters. This class also holds the meta data of the
+    current session while using the 'UNSAFE' mode option
     """
 
     def __init__(self,sessID=None, agentCount=100, numParam=10, fixParam=True,
                  spread=2, survival=0.05, genecopy=0.9, mutation=0.01, generateMode='A',mode='UNSAFE'):
-
+        
 
         self.sessID        = deamon.getNewSessionID()
         self.spread        = spread
@@ -41,23 +44,19 @@ class Session:
     def init(self):
         """
         INPUT       : None
-        OUTPUT      : Returns 1 on success, None on failure
+        OUTPUT      : None
 
-        DESCRIPTION : Initializes the session with initial
-                      values based on the session conf .
-
-        we create self.agentCount number of agents here
-        using the deamon function createAgent
+        DESCRIPTION : Initializes the session with agents
         """
         
-        if(1):
+        try:
             for i in range(self.agentCount):
                 agentObj = deamon.createAgent(self)
-        '''
-        except Exception as e:
-            print("gaDiscrete.init failed with error as : {}".format(e))
-        '''
         
+        except Exception as e:
+            print("gaDisc.init failed with error as : {}".format(e))
+
+            
     def getAllAgents(self):
         """
         INPUT       : None
@@ -65,22 +64,17 @@ class Session:
                       agents in a list, or None on error
 
         DESCRIPTION : Self explanatory from OUTPUT
-
-        
-        toRet is the set containing the agentIDs of all
-        agents who are currently active.
-        toRet is populated by using the deamon function
-        getCurrGen
         """
         
         try:
             toRet = deamon.getCurrGen(self)
         except Exception as e:
-            print("gaDiscrete.getAllAgents failed with error as : {}".format(e))
+            print("gaDisc.getAllAgents failed with error as : {}".format(e))
             toRet = None
             
         return(toRet)
 
+    
     def getAgent(self, agentID):
         """
         INPUT       : AgentID
@@ -88,10 +82,6 @@ class Session:
                       None on error
 
         DESCRIPTION : Self explanatory from OUTPUT
-
-        agentObj is the agent corresponding to the agentID = agentID
-        and belonging to the session with sessID = sessID
-        uses the deamon function
         """
 
         try:
@@ -99,26 +89,24 @@ class Session:
             return(agentObj)
         
         except Exception as e:
-            print("gaDiscrete.getAgent failed with error as : {}".format(e))
+            print("gaDisc.getAgent failed with error as : {}".format(e))
+            return(None)
+        
         
     def updateAgent(self, agentObj):
         """
         INPUT       : Agent object
         OUTPUT      : Returns 1 on success, None on failure
 
-        DESCRIPTION : Updates the status of Agent in the cluster.
-
-        Stores the agent in the current session agent's file
-        this storage is for persistance and to save the RAM
-        for tree building or other more important jobs
-        uses the deamon storeAgent function
+        DESCRIPTION : Updates the status of Agent in the current session
         """
 
         try:
             return(deamon.storeAgent( self, agentObj))
         
         except Exception as e:
-            print("gaDiscrete.updateAgent failed with error as : {}".format(e))            
+            print("gaDisc.updateAgent failed with error as : {}".format(e))
+            return(None)
     
     def getBestAgent(self):
         """
@@ -134,7 +122,7 @@ class Session:
         uses deamon functions getCurrGen, getAgent
         """
 
-        if(1):
+        try:
             
             bestAgent = None
             currGen = deamon.getCurrGen(self)
@@ -142,13 +130,14 @@ class Session:
                 agent = deamon.getAgent(self, agentID)
                 if((bestAgent is None) or (bestAgent.fitness < agent.fitness)):
                     bestAgent = agent
-        '''            
+                    
         except Exception as e:
-            print("gaDiscrete.getBestAgent failed with error as : {}".format(e))
+            print("gaDisc.getBestAgent failed with error as : {}".format(e))
             bestAgent = None
-        '''
+        
         return(bestAgent)
 
+    
     def getAverageFitness(self):
         """
         INPUT       : None
@@ -169,11 +158,12 @@ class Session:
                 totalFitness += deamon.getAgent(self, agentID).fitness
 
         except Exception:
-            print("gaDiscrete.getAverageFitness failed with error as : {}".format(e))
+            print("gaDisc.getAverageFitness failed with error as : {}".format(e))
             totalFitness = 0
 
         return(totalFitness/self.agentCount)
                 
+
     def createNextGen(self):
         """
         INPUT       : None
@@ -198,10 +188,10 @@ class Session:
         by calling a deamon function aSexualRep, which returns a
         new agentObj as a child when a parent agentObj is sent
         this new child-agentObj will inherit the fitness value and
-        slightly modified from its parent
+        slightly modified from its parent, (Needs optimisation).
         """
         
-        if(1):
+        try:
             if(self.survival > 0.95):
                 survivalPerc = 0.95
                 
@@ -230,10 +220,11 @@ class Session:
                     self.updateAgent(deamon.sexualRep(self, self.getAgent(currPop[itr]),
                                                       self.getAgent(currPop[(itr + 1)%(len(currPop))])))
                 itr = (itr + 1)%(len(currPop))
-        '''
+        
         except Exception as e:
-            print("gaDiscrete.createNextGen failed with error as : {}".format(e))
-        '''
+            print("gaDisc.createNextGen failed with error as : {}".format(e))
+
+            
     def resetEnv(self):
         """
         INPUT       : None
@@ -259,7 +250,7 @@ class Session:
             deamon.setSession(self)
             
         except Exception as e:
-            print("gaDiscrete.resetEnv failed with error as : {}".format(e))
+            print("gaDisc.resetEnv failed with error as : {}".format(e))
             
 
     def delete(self):
@@ -274,10 +265,14 @@ class Session:
         this session uses the deamon data GA_UTIL_DIR
         """
 
-        try:
-            sessDir = deamon.GA_UTIL_DIR + "/utilFiles/tmp" + str(self.sessID)
-            shutil.rmtree(sessDir)
-            del self
+        if(self.mode == 'SAFE'):
+            try:
+                sessDir = deamon.GA_UTIL_DIR + "/utilFiles/tmp" + str(self.sessID)
+                shutil.rmtree(sessDir)
+                del self
             
-        except Exception as e:
-            print("gaDiscrete.delete failed with error as : {}".format(e))
+            except Exception as e:
+                print("gaDisc.delete failed with error as : {}".format(e))
+        else:
+            del self
+
